requirements.txt
# C2 Server Dependencies
flask==3.0.2
flask-socketio==5.3.6
python-socketio==5.11.1
python-engineio==4.9.0
werkzeug==3.0.1
pyOpenSSL==24.0.0
cryptography==42.0.2
python-dotenv==1.0.1
pillow==10.2.0
pywin32==306; sys_platform == 'win32'

# Implant Dependencies
requests==2.31.0
psutil==5.9.8
pywin32==306; sys_platform == 'win32'
pillow==10.2.0
cryptography==42.0.2
python-dotenv==1.0.1 